reworked the benchmarks to not always execute node-sqlite
