+++ benchmark/trials.js
@@ -3,16 +3,10 @@
 exports.default = [
 	{ type: 'select', table: 'small', columns: ['nul', 'integer', 'real', 'text'],
 		description: 'reading rows individually' },
-	{ type: 'select-array', table: 'small', columns: ['nul', 'integer', 'real', 'text'],
-		description: 'reading rows individually into arrays' },
 	{ type: 'select-all', table: 'small', columns: ['nul', 'integer', 'real', 'text'],
 		description: 'reading 100 rows into an array' },
-	{ type: 'select-all-array', table: 'small', columns: ['nul', 'integer', 'real', 'text'],
-		description: 'reading 100 rows into an array with results as array' },
 	{ type: 'select-iterate', table: 'small', columns: ['nul', 'integer', 'real', 'text'],
 		description: 'iterating over 100 rows' },
-	{ type: 'select-iterate-array', table: 'small', columns: ['nul', 'integer', 'real', 'text'],
-		description: 'iterating over 100 rows with results as arrays' },
 	{ type: 'insert', table: 'small', columns: ['nul', 'integer', 'real', 'text'],
 		description: 'inserting rows individually' },
 	{ type: 'transaction', table: 'small', columns: ['nul', 'integer', 'real', 'text'],
@@ -27,13 +21,6 @@ exports.searchable = [
 	{ type: 'select', table: 'small', columns: ['blob'] },
 	{ type: 'select', table: 'large_text', columns: ['text'] },
 	{ type: 'select', table: 'large_blob', columns: ['blob'] },
-	{ type: 'select-array', table: 'small', columns: ['nul'] },
-	{ type: 'select-array', table: 'small', columns: ['integer'] },
-	{ type: 'select-array', table: 'small', columns: ['real'] },
-	{ type: 'select-array', table: 'small', columns: ['text'] },
-	{ type: 'select-array', table: 'small', columns: ['blob'] },
-	{ type: 'select-array', table: 'large_text', columns: ['text'] },
-	{ type: 'select-array', table: 'large_blob', columns: ['blob'] },
 	{ type: 'select-all', table: 'small', columns: ['nul'] },
 	{ type: 'select-all', table: 'small', columns: ['integer'] },
 	{ type: 'select-all', table: 'small', columns: ['real'] },
@@ -41,13 +28,6 @@ exports.searchable = [
 	{ type: 'select-all', table: 'small', columns: ['blob'] },
 	{ type: 'select-all', table: 'large_text', columns: ['text'] },
 	{ type: 'select-all', table: 'large_blob', columns: ['blob'] },
-	{ type: 'select-all-array', table: 'small', columns: ['nul'] },
-	{ type: 'select-all-array', table: 'small', columns: ['integer'] },
-	{ type: 'select-all-array', table: 'small', columns: ['real'] },
-	{ type: 'select-all-array', table: 'small', columns: ['text'] },
-	{ type: 'select-all-array', table: 'small', columns: ['blob'] },
-	{ type: 'select-all-array', table: 'large_text', columns: ['text'] },
-	{ type: 'select-all-array', table: 'large_blob', columns: ['blob'] },
 	{ type: 'select-iterate', table: 'small', columns: ['nul'] },
 	{ type: 'select-iterate', table: 'small', columns: ['integer'] },
 	{ type: 'select-iterate', table: 'small', columns: ['real'] },
@@ -55,13 +35,6 @@ exports.searchable = [
 	{ type: 'select-iterate', table: 'small', columns: ['blob'] },
 	{ type: 'select-iterate', table: 'large_text', columns: ['text'] },
 	{ type: 'select-iterate', table: 'large_blob', columns: ['blob'] },
-	{ type: 'select-iterate-array', table: 'small', columns: ['nul'] },
-	{ type: 'select-iterate-array', table: 'small', columns: ['integer'] },
-	{ type: 'select-iterate-array', table: 'small', columns: ['real'] },
-	{ type: 'select-iterate-array', table: 'small', columns: ['text'] },
-	{ type: 'select-iterate-array', table: 'small', columns: ['blob'] },
-	{ type: 'select-iterate-array', table: 'large_text', columns: ['text'] },
-	{ type: 'select-iterate-array', table: 'large_blob', columns: ['blob'] },
 	{ type: 'insert', table: 'small', columns: ['nul'] },
 	{ type: 'insert', table: 'small', columns: ['integer'] },
 	{ type: 'insert', table: 'small', columns: ['real'] },
