+++ benchmark/types/select-all.js
@@ -7,6 +7,13 @@ exports['better-sqlite3'] = (db, { table, columns, count }) => {
 	return () => stmt.all((rowid += 100) % count + 1);
 };
 
+exports['better-sqlite3-array'] = (db, { table, columns, count }) => {
+	const stmt = db.prepare(`SELECT ${columns.join(', ')} FROM ${table} WHERE rowid >= ? LIMIT 100`);
+	stmt.asArray();
+	let rowid = -100;
+	return () => stmt.all((rowid += 100) % count + 1);
+};
+
 exports['node-sqlite3'] = async (db, { table, columns, count }) => {
 	const sql = `SELECT ${columns.join(', ')} FROM ${table} WHERE rowid >= ? LIMIT 100`;
 	let rowid = -100;
