+++ benchmark/types/select.js
@@ -7,6 +7,13 @@ exports['better-sqlite3'] = (db, { table, columns, count }) => {
 	return () => stmt.get(++rowid % count + 1);
 };
 
+exports['better-sqlite3-array'] = (db, { table, columns, count }) => {
+	const stmt = db.prepare(`SELECT ${columns.join(', ')} FROM ${table} WHERE rowid = ?`);
+	stmt.asArray();
+	let rowid = -1;
+	return () => stmt.get(++rowid % count + 1);
+};
+
 exports['node-sqlite3'] = async (db, { table, columns, count }) => {
 	const sql = `SELECT ${columns.join(', ')} FROM ${table} WHERE rowid = ?`;
 	let rowid = -1;
